pyproject.toml
[project]
name = "livros-pel"
version = "0.1.0"
description = ""
authors = [
    {name = "Iago Cruz",email = "iagosilvacontato@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "django[standard] (>=5.2.7,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg[binary] (>=3.2.12,<4.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.14.3,<0.15.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "freezegun (>=1.5.5,<2.0.0)"
]

[tool.taskipy.tasks]
lint = 'ruff check .'
pre_format = 'ruff check --fix .'
format = 'ruff format .'
run = 'python manage.py runserver'
migrate = 'python manage.py migrate'
makemigrations = 'python manage.py makemigrations'
pre_test = 'task lint'
test = 'pytest -s -x --cov=. -vv'
shell = 'python manage.py shell'